[#20] refactor : 폴더구조 정리
