+++ grapit-spring/src/main/java/edu/oak/grapitspring/common/rtc/ViduController.java
@@ -1,4 +1,4 @@
-package edu.oak.grapitspring.rtc;
+package edu.oak.grapitspring.common.rtc;
 
 
 import io.openvidu.java.client.*;
