+++ grapit-spring/src/main/java/edu/oak/grapitspring/controller/Room/CreateRoomRequestDTO.java
@@ -1,4 +1,4 @@
-package edu.oak.grapitspring.controller;
+package edu.oak.grapitspring.controller.Room;
 
 import edu.oak.grapitspring.domain.Room;
 import lombok.*;
