+++ grapit-spring/src/main/java/edu/oak/grapitspring/service/Room/RoomResponseDTO.java
@@ -1,4 +1,4 @@
-package edu.oak.grapitspring.service;
+package edu.oak.grapitspring.service.Room;
 
 import lombok.Builder;
 import lombok.Data;
